chore: converted ply read/writer to stl format
